Add StartsWithAny and EndsWithAny
