--- source/Nuke.Common/Utilities/String.OrdinalIgnoreCase.cs
@@ -1,37 +1,0 @@
-// Copyright 2019 Maintainers of NUKE.
-// Distributed under the MIT License.
-// https://github.com/nuke-build/nuke/blob/master/LICENSE
-
-using System;
-using System.Linq;
-using JetBrains.Annotations;
-
-namespace Nuke.Common.Utilities
-{
-    public static partial class StringExtensions
-    {
-        [Pure]
-        public static bool ContainsOrdinalIgnoreCase(this string str, string other)
-        {
-            return str.IndexOf(other, StringComparison.OrdinalIgnoreCase) >= 0;
-        }
-
-        [Pure]
-        public static bool EqualsOrdinalIgnoreCase(this string str, string other)
-        {
-            return str.Equals(other, StringComparison.OrdinalIgnoreCase);
-        }
-
-        [Pure]
-        public static bool StartsWithOrdinalIgnoreCase(this string str, string other)
-        {
-            return str.StartsWith(other, StringComparison.OrdinalIgnoreCase);
-        }
-
-        [Pure]
-        public static bool EndsWithOrdinalIgnoreCase(this string str, string other)
-        {
-            return str.EndsWith(other, StringComparison.OrdinalIgnoreCase);
-        }
-    }
-}
